clean up

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
  */
 int main(void)
 {
-	unsigned long int n, n_1, next;
+	unsigned long long int n, n_1, next;
 	int i;
 
 	n = 2;
 	n_1 = 1;
-	printf("%lu, ", n_1);
+	printf("%llu, ", n_1);
 	for (i = 0; i < 98; i++)
 	{
-		printf("%lu", n);
+		printf("%llu", n);
 		if (i != 97)
 			printf(", ");
 		next = n + n_1;
